requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aioredis==1.3.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
attrs==23.2.0
Authlib==1.3.0
bcrypt==4.1.2
blinker==1.8.2
boto3==1.34.69
botocore==1.34.69
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
decorator==5.1.1
Deprecated==1.2.14
dnspython==2.6.1
ecdsa==0.18.0
elastic-transport==8.13.1
elasticsearch==8.14.0
elasticsearch-dsl==8.15.0
email_validator==2.1.1
fastapi==0.110.0
fastapi-cache==0.1.0
fastapi-limiter==0.1.6
filelock==3.15.4
Flask==3.0.3
frozenlist==1.4.1
fsspec==2024.6.1
future==1.0.0
geocoder==1.38.1
geographiclib==2.0
geopy==2.4.1
googlemaps==4.10.0
graphql-core==3.2.3
grpcio==1.64.0
grpcio-tools==1.64.0
h11==0.14.0
haversine==2.8.1
hiredis==2.3.2
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
idna==3.6
importlib_resources==6.4.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.2
limits==3.10.1
MarkupSafe==2.1.5
motor==3.3.2
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
numpy==1.26.0
openrouteservice==2.3.3
orjson==3.10.6
packaging==24.0
passlib==1.7.4
paypalrestsdk==1.13.3
pillow==10.2.0
pluggy==1.4.0
protobuf==5.27.0
pyasn1==0.5.1
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
pymongo==4.6.2
pyOpenSSL==24.1.0
pytest==8.1.1
pytest-asyncio==0.23.7
pytest-dotenv==0.5.2
pytest-env==1.1.3
pytest-flask==1.3.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
ratelim==0.1.6
redis==5.1.0b4
regex==2024.7.24
requests==2.31.0
rsa==4.9
s3transfer==0.10.1
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.0
sendgrid==6.11.0
sentence-transformers==3.0.1
setuptools==70.0.0
six==1.16.0
slowapi==0.1.9
sniffio==1.3.1
starkbank-ecdsa==2.2.0
starlette==0.36.3
strawberry-graphql==0.220.0
stripe==9.7.0
sympy==1.13.2
tenacity==8.4.2
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.0
tqdm==4.66.5
transformers==4.44.0
twilio==9.0.2
typing_extensions==4.10.0
unicorn==2.0.1.post1
urllib3==2.2.1
uvicorn==0.28.0
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
